feat(library/compiler): clean up native compiler passes, and remove flatten_let
