+++ src/library/compiler/CMakeLists.txt
@@ -4,4 +4,4 @@ add_library(compiler OBJECT util.cpp eta_expansion.cpp simp_pr1_rec.cpp preproce
   lambda_lifting.cpp simp_inductive.cpp nat_value.cpp
   vm_compiler.cpp init_module.cpp native_compiler.cpp
   cpp_emitter.cpp config.cpp used_names.cpp annotate_return.cpp
-  anf_transform.cpp flatten_let.cpp)
+  anf_transform.cpp)
